pyproject.toml
[tool.poetry]
name = "ai-coder"
version = "0.1.0"
description = "An experimental code generation tool powered by OpenAI's ChatGPT"
authors = ["Pascal Cremer <pascal.cremer@geekcell.io>"]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.173"
openai = "^0.27.6"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
